+++ src/models/Settings.php
@@ -1,31 +1,11 @@
 <?php
-/**
- * craft-ccb-login plugin for Craft CMS 3.x
- *
- * Log in via the CCB API, and log user info in session.
- *
- * @link      thisanimus.com
- * @copyright Copyright (c) 2019 Andrew Hale
- */
 
-namespace thisanimus\craftccblogin\models;
+namespace thisanimus\CCBLogin\models;
 
-use thisanimus\craftccblogin\Craftccblogin;
-
-use Craft;
 use craft\base\Model;
+use craft\behaviors\EnvAttributeParserBehavior;
 
-/**
- * Craftccblogin Settings Model
- *
- * https://craftcms.com/docs/plugins/models
- *
- * @author    Andrew Hale
- * @package   Craftccblogin
- * @since     1.0.11
- */
-class Settings extends Model
-{
+class Settings extends Model{
     /**
      *
      * @var string